# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "wday3",
        "description": "Backend sistema de gestión de personal",
        "contact": {
            "email": "totomarion@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/registrarse": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Crear de usuario",
                "operationId": "registrarseAuthController",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRegistrarseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Login de usuario",
                "operationId": "loginAuthController",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario logueado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Logout de usuario",
                "operationId": "logoutAuthController",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado": {
            "put": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Modificar un empleado",
                "operationId": "empleadoUpdate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado modificado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Crear empleado",
                "operationId": "storeEmpleadoController",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado/{id}": {
            "get": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Buscar un empleado",
                "operationId": "empleadoShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "EmpleadoController"
                ],
                "summary": "Eliminar un empleado",
                "operationId": "empleadoDestroy",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpleadoDestroyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/create": {
            "post": {
                "tags": [
                    "PasswordResetController"
                ],
                "summary": "Solicitar reset contraseña",
                "operationId": "createPasswordResetController",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Solicitud no válida"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error validación"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthLoginRequest": {
                "title": "AuthLoginRequest",
                "description": "Registrarse Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "AuthRegistrarseRequest": {
                "title": "AuthRegistrarseRequest",
                "description": "Registrarse Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "EmpleadoDestroyRequest": {
                "title": "EmpleadoDestroyRequest",
                "description": "Empleado Destroy Request",
                "properties": {
                    "id": {
                        "description": "Id del empleado",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "EmpleadoStoreRequest": {
                "title": "EmpleadoStoreRequest",
                "description": "Empleado Store Request",
                "properties": {
                    "user_id": {
                        "description": "Id del usuario",
                        "type": "integer"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    },
                    "nombre": {
                        "description": "Nombre del empleado",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email del empleado",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "EmpleadoUpdateRequest": {
                "title": "EmpleadoUpdateRequest",
                "description": "Empleado Update Request",
                "properties": {
                    "id": {
                        "description": "Id del empleado",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del empleado",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PasswordResetCreateRequest": {
                "title": "PasswordResetCreateRequest",
                "description": "Password Reset Request",
                "properties": {
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "Empleado": {
                "title": "Empleado",
                "description": "Empleado representation",
                "properties": {
                    "id": {
                        "description": "Id del empleado",
                        "type": "integer"
                    },
                    "apellido": {
                        "description": "Apellido del empleado",
                        "type": "string"
                    },
                    "nombre": {
                        "description": "Nombre del empleado",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Representación del usuario",
                "properties": {
                    "id": {
                        "description": "Id del usuario",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Cuando el usuario verifica su email",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "active": {
                        "description": "Si usuario se encuentra activo",
                        "type": "boolean"
                    },
                    "empleados": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Empleado"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}